+++ tests/unit/general/test_oauth.py
@@ -302,7 +302,7 @@ def test_oauth_config_wrong_relation_order(
         remote_app_data={**OAUTH_RELATION_DATA_EXAMPLE, "client_secret_id": secret_id},
     )
     base_state["relations"].append(oauth_relation)
-    base_state["secrets"] = [testing.Secret(id=secret_id, tracked_content={"secret": "abc"})]
+    base_state["secrets"].append(testing.Secret(id=secret_id, tracked_content={"secret": "abc"}))
 
     state = testing.State(**base_state)
     context = context_factory(charm)
@@ -622,7 +622,7 @@ def test_oauth_config_correct_relation_order(
         remote_app_data={**OAUTH_RELATION_DATA_EXAMPLE, "client_secret_id": secret_id},
     )
     base_state["relations"].append(oauth_relation)
-    base_state["secrets"] = [testing.Secret(id=secret_id, tracked_content={"secret": "abc"})]
+    base_state["secrets"].append(testing.Secret(id=secret_id, tracked_content={"secret": "abc"}))
 
     state = testing.State(**base_state)
     out = context.run(context.on.relation_changed(oauth_relation), state)
@@ -723,7 +723,7 @@ def test_oauth_config_remove_ingress_integration_should_block(
         remote_app_data={**OAUTH_RELATION_DATA_EXAMPLE, "client_secret_id": secret_id},
     )
     base_state["relations"].append(oauth_relation)
-    base_state["secrets"] = [testing.Secret(id=secret_id, tracked_content={"secret": "abc"})]
+    base_state["secrets"].append(testing.Secret(id=secret_id, tracked_content={"secret": "abc"}))
 
     state = testing.State(**base_state)
     out = context.run(context.on.relation_changed(oauth_relation), state)
@@ -825,7 +825,7 @@ def test_oauth_config_remove_oauth_integration_should_not_block(
         remote_app_data={**OAUTH_RELATION_DATA_EXAMPLE, "client_secret_id": secret_id},
     )
     base_state["relations"].append(oauth_relation)
-    base_state["secrets"] = [testing.Secret(id=secret_id, tracked_content={"secret": "abc"})]
+    base_state["secrets"].append(testing.Secret(id=secret_id, tracked_content={"secret": "abc"}))
 
     state = testing.State(**base_state)
     out = context.run(context.on.relation_changed(oauth_relation), state)
@@ -915,7 +915,7 @@ def test_oauth_config_wrong_scope(
         remote_app_data={**OAUTH_RELATION_DATA_EXAMPLE, "client_secret_id": secret_id},
     )
     base_state["relations"].append(oauth_relation)
-    base_state["secrets"] = [testing.Secret(id=secret_id, tracked_content={"secret": "abc"})]
+    base_state["secrets"].append(testing.Secret(id=secret_id, tracked_content={"secret": "abc"}))
     ingress_relation = testing.Relation(
         endpoint="ingress",
         interface="ingress",
@@ -1006,7 +1006,7 @@ def test_blocked_when_relation_data_empty(
         endpoint="oidc", interface="oauth", remote_app_data={}, remote_app_name="OIDC_CHARM"
     )
     base_state["relations"].append(oauth_relation)
-    base_state["secrets"] = [testing.Secret(id=secret_id, tracked_content={"secret": "abc"})]
+    base_state["secrets"].append(testing.Secret(id=secret_id, tracked_content={"secret": "abc"}))
     ingress_relation = testing.Relation(
         endpoint="ingress",
         interface="ingress",
@@ -1121,7 +1121,7 @@ def test_oauth_multiple_oauth_integrations(
         remote_app_data={**OAUTH_RELATION_DATA_EXAMPLE, "client_secret_id": secret_id},
     )
     base_state["relations"].append(oauth_relation)
-    base_state["secrets"] = [testing.Secret(id=secret_id, tracked_content={"secret": "abc"})]
+    base_state["secrets"].append(testing.Secret(id=secret_id, tracked_content={"secret": "abc"}))
     ingress_relation = testing.Relation(
         endpoint="ingress",
         interface="ingress",